requirements.txt
Django==5.2
django-filter==25.1
requests==2.32.3
dotenv
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
sympy==1.13.3
djangorestframework-api-key==3.1.0
celery==5.5.0
django-celery-beat==2.8.0
django-redis==5.4.0
django-extensions==4.1
django-debug-toolbar==5.1.0
drf-spectacular==0.28.0
drf-writable-nested==0.7.2
drf_link_header_pagination==0.2.0
psycopg
psycopg-binary
django-environ==0.12.0
twilio==9.6.2



#for PostgreSQL
# psycopg2-binary is a precompiled version of psycopg2, which is a PostgreSQL adapter for Python.
# psycopg2-binary==2.9.10
# psycopg2==2.9. MapSerializer
